fix fishing-map circular dependencies
