+++ apps/fishing-map/features/datasets/NewDatasetTooltip.tsx
@@ -7,7 +7,8 @@ import { Spinner } from '@globalfishingwatch/ui-components'
 import { Dataset, DatasetCategory } from '@globalfishingwatch/api-types'
 import LocalStorageLoginLink from 'routes/LoginLink'
 import { AsyncReducerStatus } from 'utils/async-slice'
-import { isGuestUser, selectUserDatasetsNotUsed } from 'features/user/user.selectors'
+import { selectUserDatasetsNotUsed } from 'features/user/user.selectors'
+import { isGuestUser } from 'features/selectors/user.selectors'
 import { useDatasetModalConnect, useAddDataviewFromDatasetToWorkspace } from './datasets.hook'
 import styles from './NewDatasetTooltip.module.css'
 import {
