+++ apps/fishing-map/features/editor/WorkspaceEditor.tsx
@@ -5,7 +5,7 @@ import { groupBy } from 'lodash'
 import { Dataview, DataviewCategory } from '@globalfishingwatch/api-types'
 import { Spinner, IconButton } from '@globalfishingwatch/ui-components'
 import { AsyncError, AsyncReducerStatus } from 'utils/async-slice'
-import { selectDataviewInstancesMerged } from 'features/dataviews/dataviews.selectors'
+import { selectDataviewInstancesMerged } from 'features/selectors/dataviews.selectors'
 import { useDataviewInstancesConnect } from 'features/workspace/workspace.hook'
 import { addDataviewEntity } from 'features/dataviews/dataviews.slice'
 import { getDataviewInstanceFromDataview } from 'features/dataviews/dataviews.utils'
