+++ apps/fishing-map/features/feedback/FeedbackModal.tsx
@@ -7,8 +7,9 @@ import { selectActiveDataviews } from 'features/dataviews/dataviews.selectors'
 import { getSourcesSelectedInDataview } from 'features/workspace/activity/activity.utils'
 import { GUEST_USER_TYPE, selectUserData } from 'features/user/user.slice'
 import { loadSpreadsheetDoc } from 'utils/spreadsheet'
-import { isGuestUser, selectUserGroupsClean } from 'features/user/user.selectors'
+import { selectUserGroupsClean } from 'features/user/user.selectors'
 import { getDatasetLabel } from 'features/datasets/datasets.utils'
+import { isGuestUser } from 'features/selectors/user.selectors'
 import styles from './FeedbackModal.module.css'
 
 type FeedbackModalProps = {
