+++ apps/fishing-map/features/map/controls/MapControls.tsx
@@ -15,7 +15,7 @@ import { BasemapType, GeneratorType } from '@globalfishingwatch/layer-composer'
 import { getOceanAreaName, OceanAreaLocale } from '@globalfishingwatch/ocean-areas'
 import { useDebounce } from '@globalfishingwatch/react-hooks'
 import { useDataviewInstancesConnect } from 'features/workspace/workspace.hook'
-import { selectDataviewInstancesResolved } from 'features/dataviews/dataviews.selectors'
+import { selectDataviewInstancesResolved } from 'features/selectors/dataviews.selectors'
 import Rulers from 'features/map/rulers/Rulers'
 import useViewport, { useMapBounds } from 'features/map/map-viewport.hooks'
 import { isWorkspaceLocation } from 'routes/routes.selectors'
