+++ apps/fishing-map/features/map/map-features.atom.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil'
+
+export const mapIdleAtom = atom<boolean>({
+  key: 'mapIdle',
+  default: false,
+})
