+++ apps/fishing-map/features/map/map-features.hooks.ts
@@ -1,12 +1,8 @@
 import { useEffect } from 'react'
-import { atom, useRecoilValue, useSetRecoilState } from 'recoil'
+import { useRecoilValue, useSetRecoilState } from 'recoil'
 import useMapInstance from 'features/map/map-context.hooks'
-import { useMapStyle } from './map.hooks'
-
-export const mapIdleAtom = atom<boolean>({
-  key: 'mapIdle',
-  default: false,
-})
+import { useMapLoaded, useMapStyle } from 'features/map/map-style.hooks'
+import { mapIdleAtom } from 'features/map/map-features.atom'
 
 export const useSetMapIdleAtom = () => {
   // Used it once in Map.tsx the listeners only once
@@ -37,17 +33,6 @@ export const useSetMapIdleAtom = () => {
 export const useMapIdle = () => {
   const idle = useRecoilValue(mapIdleAtom)
   return idle
-}
-
-export const useMapLoaded = () => {
-  const idle = useRecoilValue(mapIdleAtom)
-  const map = useMapInstance()
-  const mapInstanceReady = map !== null
-  const mapFirstLoad = (map as any)?._loaded || false
-  const mapStyleLoad = map?.isStyleLoaded() || false
-  const areTilesLoaded = map?.areTilesLoaded() || false
-  const loaded = mapInstanceReady && mapFirstLoad && (idle || areTilesLoaded || mapStyleLoad)
-  return loaded
 }
 
 export const useSourceInStyle = (sourcesIds: string | string[]) => {
