+++ apps/fishing-map/features/map/map-viewport.hooks.ts
@@ -9,9 +9,9 @@ import { LngLatBounds } from '@globalfishingwatch/mapbox-gl'
 import { Bbox, MapCoordinates } from 'types'
 import { DEFAULT_VIEWPORT } from 'data/config'
 import { updateUrlViewport } from 'routes/routes.actions'
-import { TIMEBAR_HEIGHT } from 'features/timebar/Timebar'
 import { FOOTER_HEIGHT } from 'features/footer/Footer'
 import { selectViewport } from 'features/app/app.selectors'
+import { TIMEBAR_HEIGHT } from 'features/timebar/timebar.config'
 import store, { RootState } from '../../store'
 import useMapInstance from './map-context.hooks'
