+++ apps/fishing-map/features/resources/resources.selectors.ts
@@ -1,8 +1,5 @@
 import { createSelector } from '@reduxjs/toolkit'
 import { selectVisibleEvents } from 'features/app/app.selectors'
-import { ThinningLevels, THINNING_LEVELS } from 'data/config'
-import { selectDebugOptions } from 'features/debug/debug.slice'
-import { isGuestUser } from 'features/user/user.selectors'
 import { RootState } from 'store'
 import { selectResources } from './resources.slice'
 
@@ -19,16 +16,5 @@ export const selectVisibleResources = createSelector(
           : true
       })
     )
-  }
-)
-
-export const selectThinningConfig = createSelector(
-  [(state) => isGuestUser(state), selectDebugOptions],
-  (guestUser, { thinning }) => {
-    if (!thinning) return null
-    const thinningConfig = guestUser
-      ? THINNING_LEVELS[ThinningLevels.Aggressive]
-      : THINNING_LEVELS[ThinningLevels.Default]
-    return thinningConfig
   }
 )
