+++ apps/fishing-map/features/search/Search.tsx
@@ -29,7 +29,7 @@ import I18nDate from 'features/i18n/i18nDate'
 import LocalStorageLoginLink from 'routes/LoginLink'
 import { AsyncReducerStatus } from 'utils/async-slice'
 import { EMPTY_FIELD_PLACEHOLDER, formatInfoField } from 'utils/info'
-import { selectVesselsDataviews } from 'features/dataviews/dataviews.selectors'
+import { selectVesselsDataviews } from 'features/selectors/dataviews.selectors'
 import I18nFlag from 'features/i18n/i18nFlag'
 import { FIRST_YEAR_OF_DATA } from 'data/config'
 import {
