+++ apps/fishing-map/features/search/search.selectors.ts
@@ -2,7 +2,7 @@ import { createSelector } from '@reduxjs/toolkit'
 import { checkExistPermissionInList } from 'auth-middleware/src/utils'
 import { selectUserData } from 'features/user/user.slice'
 import { selectVesselsDatasets } from 'features/datasets/datasets.selectors'
-import { isGuestUser } from 'features/user/user.selectors'
+import { isGuestUser } from 'features/selectors/user.selectors'
 import { filterDatasetsByUserType, getDatasetsInDataviews } from 'features/datasets/datasets.utils'
 import { selectAllDataviewsInWorkspace } from 'features/dataviews/dataviews.selectors'
 import { selectAllDatasets } from 'features/datasets/datasets.slice'
