+++ apps/fishing-map/features/sidebar/CategoryTabs.tsx
@@ -9,7 +9,7 @@ import { WorkspaceCategories } from 'data/workspaces'
 import { HOME, USER, WORKSPACES_LIST } from 'routes/routes'
 import { selectLocationCategory, selectLocationType } from 'routes/routes.selectors'
 import { selectUserData } from 'features/user/user.slice'
-import { isGuestUser } from 'features/user/user.selectors'
+import { isGuestUser } from 'features/selectors/user.selectors'
 import { useClickedEventConnect } from 'features/map/map.hooks'
 import useMapInstance from 'features/map/map-context.hooks'
 import { selectAvailableWorkspacesCategories } from 'features/workspaces-list/workspaces-list.selectors'
