+++ apps/fishing-map/features/sidebar/SidebarHeader.tsx
@@ -15,7 +15,7 @@ import { AsyncReducerStatus } from 'utils/async-slice'
 import { isWorkspaceLocation, selectLocationCategory } from 'routes/routes.selectors'
 import { WorkspaceCategories } from 'data/workspaces'
 import { selectWorkspaceWithCurrentState, selectReadOnly } from 'features/app/app.selectors'
-import { isGuestUser } from 'features/user/user.selectors'
+import { isGuestUser } from 'features/selectors/user.selectors'
 import NewWorkspaceModal from 'features/workspace/shared/NewWorkspaceModal'
 import { useClipboardNotification } from './sidebar.hooks'
 import styles from './SidebarHeader.module.css'
