+++ apps/fishing-map/features/timebar/timebar.config.ts
@@ -1,0 +1,2 @@
+export const TIMEBAR_HEIGHT = 72
+export const MAX_TIMEBAR_VESSELS = 10
