+++ apps/fishing-map/features/workspace/Workspace.tsx
@@ -10,7 +10,7 @@ import {
 } from 'features/workspace/workspace.selectors'
 import { fetchResourceThunk } from 'features/resources/resources.slice'
 import { AsyncReducerStatus } from 'utils/async-slice'
-import { isGuestUser } from 'features/user/user.selectors'
+import { isGuestUser } from 'features/selectors/user.selectors'
 import { selectLocationCategory, selectWorkspaceId } from 'routes/routes.selectors'
 import { HOME } from 'routes/routes'
 import { updateLocation } from 'routes/routes.actions'
