+++ apps/fishing-map/features/workspace/environmental/EnvironmentalLayerPanel.tsx
@@ -7,8 +7,9 @@ import { Tooltip, ColorBarOption } from '@globalfishingwatch/ui-components'
 import { UrlDataviewInstance } from '@globalfishingwatch/dataviews-client'
 import styles from 'features/workspace/shared/LayerPanel.module.css'
 import { useDataviewInstancesConnect } from 'features/workspace/workspace.hook'
-import { isGuestUser, selectUserId } from 'features/user/user.selectors'
+import { selectUserId } from 'features/user/user.selectors'
 import { useAutoRefreshImportingDataset } from 'features/datasets/datasets.hook'
+import { isGuestUser } from 'features/selectors/user.selectors'
 import DatasetNotFound from '../shared/DatasetNotFound'
 import Color from '../common/Color'
 import LayerSwitch from '../common/LayerSwitch'
