+++ apps/fishing-map/features/workspace/workspace.hook.ts
@@ -2,10 +2,14 @@ import { useSelector } from 'react-redux'
 import { useCallback } from 'react'
 import { UrlDataviewInstance } from '@globalfishingwatch/dataviews-client'
 import { ColorCyclingType } from '@globalfishingwatch/api-types'
-import { FillColorBarOptions, LineColorBarOptions , ColorBarOption } from '@globalfishingwatch/ui-components'
+import {
+  FillColorBarOptions,
+  LineColorBarOptions,
+  ColorBarOption,
+} from '@globalfishingwatch/ui-components'
 import { selectUrlDataviewInstances } from 'routes/routes.selectors'
 import { useLocationConnect } from 'routes/routes.hook'
-import { selectDataviewInstancesResolved } from 'features/dataviews/dataviews.selectors'
+import { selectDataviewInstancesResolved } from 'features/selectors/dataviews.selectors'
 import { selectWorkspaceDataviewInstances } from './workspace.selectors'
 
 const getNextColor = (
