+++ apps/fishing-map/features/workspace/workspace.slice.ts
@@ -25,9 +25,9 @@ import {
 } from 'data/workspaces'
 import { AsyncReducerStatus, AsyncError } from 'utils/async-slice'
 import { getDatasetsInDataviews } from 'features/datasets/datasets.utils'
-import { isGuestUser } from 'features/user/user.selectors'
 import { isGFWUser } from 'features/user/user.slice'
 import { AppWorkspace } from 'features/workspaces-list/workspaces-list.slice'
+import { isGuestUser } from 'features/selectors/user.selectors'
 import { selectWorkspaceStatus } from './workspace.selectors'
 
 type LastWorkspaceVisited = { type: string; payload: any; query: any }
